+++ src/rex.forms/NEWS.rst
@@ -14,6 +14,7 @@ REX.FORMS Change Log
 - All get_by_uid() and find() methods now accept and optional user parameter to
   indicate that the resulting instance should be accessible by the specified
   User.
+- Fixed rendering of boolean fields as dropDown widgets.
 
 
 0.21.0 (7/31/2014)